refactor: round 1
