+++ src/defaults/config.json
@@ -1,0 +1,28 @@
+{
+  "dimensions": {
+      "display_height":480,
+      "display_width":640,
+      "processing_height": 820,
+      "processing_width": 666,
+      "windowWidth": 1280,
+      "windowHeight": 720
+  },
+  "threshold_params":{
+    "GAMMA_LOW": 0.7,
+    "MIN_GAP": 30,
+    "MIN_JUMP": 25,
+    "JUMP_DELTA": 30
+  },
+  "alignment_params":{
+    "match_col": 5,
+    "max_steps": 20,
+    "stride": 1,
+    "thickness": 3
+  },
+  "outputs":{
+    "show_image_level":4,
+    "save_image_level":2,
+    "save_detections":true,
+    "explain_scoring":0
+  }
+}
